Duplicate Ticket Id Issue

+++ datadump.sql
@@ -1,6 +1,6 @@
 -- This query generates the data for the contacts table
-drop table gettysburgstaging..contacts
-
+IF EXISTS (SELECT * FROM gettysburgstaging.INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = N'contacts')
+  drop table gettysburgstaging..contacts
 select cltcode, cltfirstname, cltSurName,cltEmail into gettysburgstaging..contacts
 from clients where cltcode in 
 (select distinct timailinglist from tickets 
@@ -11,20 +11,8 @@ inner join till on sftill = tilCode
 left outer join clients on timailinglist = cltcode
 where eveventdate between '2014-10-01' and '2015-01-31 23:59')
 
-/*
--- This query generates the data for the users table
-select tilcode userid, tildescr username from till where tilcode in
-(select distinct tilcode from tickets 
-inner join events on tievent = evcode
-inner join shows on evshow = shcode
-inner join shifts on tishift = sfcode and sfActionType = 0
-inner join till on sftill = tilCode
-left outer join clients on timailinglist = cltcode
-where eveventdate between '2014-10-01' and '2015-01-31 23:59')
 
--- This query generates the data for the guides table
-select gdrcode+1000 guideid,gdrfirstname + ' ' + gdrlastname guidename from guiders
-*/
+
 
 drop table gettysburgstaging..pricetypes 
 -- This query generates the data for the pricetype table
@@ -68,10 +56,11 @@ where eveventdate between '2014-10-01' and '2015-01-31 23:59'
 update t1
 set minticode = minticode2
 from 
-temptickets t1 inner join (select min(ticode) minticode2, tipricetype,tifullprice,eveventdate, shcode, tilcode, timailinglist
-from temptickets group by tipricetype,tifullprice,eveventdate, shcode, tilcode, timailinglist) t2 on
+temptickets t1 inner join (select min(ticode) minticode2, tipricetype,tifullprice,eveventdate, shcode, tilcode, timailinglist,tistatus,guideid,ptdescr,rcbshiftdate
+from temptickets group by tipricetype,tifullprice,eveventdate, shcode, tilcode, timailinglist,tistatus,guideid,ptdescr,rcbshiftdate) t2 on
 t1.tipricetype = t2.tipricetype and t1.tifullprice = t2.tifullprice and t1.eveventdate = t2.eveventdate and
-t1.shcode = t2.shcode and t1.tilcode = t2.tilcode and t1.timailinglist = t2.timailinglist
+t1.shcode = t2.shcode and t1.tilcode = t2.tilcode and t1.timailinglist = t2.timailinglist and t1.tistatus = t2.tistatus and isnull(t1.guideid,0) = isnull(t2.guideid,0)
+and isnull(t1.ptdescr,'') = isnull(t2.ptdescr,'') and isnull(t1.rcbshiftdate,'2000-01-01') = isnull(t2.rcbshiftdate,'2000-01-01')
 
 update t1
 set paymentid = paymentid2
@@ -85,7 +74,7 @@ t1.timailinglist = t2.timailinglist and t1.ptdescr = t2.ptdescr and t1.rcbshiftd
 drop table GettysburgStaging..tickets
 
 select minticode tickid, tipricetype pricetypeid, tiFullPrice price,evEventDate,shcode Activity,tilcode userid,tiMailingList contactid,count(*) qty, 
-case when tistatus = 9 then 'Return' else 'Sale' end status, guideid, fees,case when paymentid = 0 then null else paymentid end paymentid, max(titransactnum) last_transact_no 
+case when tistatus = 9 then 'Return' else 'Sale' end status, guideid+1000 guideid, fees,case when paymentid = 0 then null else paymentid end paymentid, max(titransactnum) last_transact_no 
 into GettysburgStaging..tickets
 from temptickets
 group by minticode, tipricetype, tiFullPrice,evEventDate,shcode,tilcode,tiMailingList,tistatus,guideid, fees,paymentid 
@@ -105,15 +94,6 @@ where paymentid > 0
 group by paymentid,timailinglist,ptdescr,rcbshiftdate, rcbcachinout
 
 
-drop table gettysburgstaging..users
-go
-
-select user_name + '@gettysburgfoundation' guideid,first_name, last_name, cltemail email,pecode externalid into gettysburgstaging..users
-from SugarGettysburg..users u 
-inner join pegettysburg..ASC_SoapSyncCrossReference on u.id = sugarid and sugarmodule = 'Users'
-inner join pegettysburg..Guiders on gdrCode = pecode
-INNER JOIN pegettysburg..Clients ON gdrContactNoInRelTable = cltCode
-where pecode in (select guideid from GettysburgStaging..tickets)
 /* This is the old versions
 -- This query generates the data for the tickets table
 select min(ticode) tickid, tipricetype pricetypeid, tiFullPrice price,evEventDate,shcode Activity,tilcode userid,tiMailingList contactid,count(*) qty,
@@ -149,6 +129,8 @@ from events inner join shows on evshow = shcode
 group by evshow,convert(varchar(5),eveventdate,8),convert(varchar(5),dateadd(n,shlongminutes,eveventdate),8)
 
 */
+IF EXISTS (SELECT * FROM gettysburgstaging.INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = N'activitypriceschedule')
+drop table gettysburgstaging..activitypriceschedule
 select min(evcode) activityscheduleid,pctcode pricetypeid,min(pdcalculatedprice) price into gettysburgstaging..activitypriceschedule
 FROM PriceType 
 INNER JOIN PriceDiscount ON PriceType.pctCode = PriceDiscount.pdPriceType 
@@ -164,3 +146,29 @@ inner join till on sftill = tilCode
 left outer join clients on timailinglist = cltcode
 where eveventdate between '2014-10-01' and '2015-01-31 23:59')
 group by evshow,convert(varchar(5),eveventdate,8),convert(varchar(5),dateadd(n,shlongminutes,eveventdate),8),pctcode 
+
+
+IF EXISTS (SELECT * FROM gettysburgstaging.INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = N'users')
+drop table gettysburgstaging..users
+go
+
+-- This query generates the data for the users table
+
+select user_name + '@gettysburgfoundation' username,first_name, last_name, cltemail email,pecode + 1000 externalid into gettysburgstaging..users
+from SugarGettysburg..users u 
+inner join pegettysburg2..ASC_SoapSyncCrossReference on u.id = sugarid and sugarmodule = 'Users'
+inner join pegettysburg2..Guiders on gdrCode = pecode
+INNER JOIN pegettysburg2..Clients ON gdrContactNoInRelTable = cltCode
+where pecode in (select guideid -1000 guideid from GettysburgStaging..tickets)
+
+insert into gettysburgstaging..users (username,first_name, last_name, email,externalid)
+
+select replace(tildescr,' ','_') + 'peuser@gettysburgfoundation','',replace(tildescr,' ','_'),'',tilcode externalid from till where tilcode in
+(select distinct tilcode from tickets 
+inner join events on tievent = evcode
+inner join shows on evshow = shcode
+inner join shifts on tishift = sfcode and sfActionType = 0
+inner join till on sftill = tilCode
+left outer join clients on timailinglist = cltcode
+where eveventdate between '2014-10-01' and '2015-01-31 23:59')
+
